move observer to UI
