+++ algobattle/match.py
@@ -6,7 +6,7 @@ from typing import Any, Self
 from prettytable import PrettyTable, DOUBLE_BORDER
 
 from algobattle.battle_wrapper import BattleWrapper, Iterated
-from algobattle.observer import Observer, Subject
+from algobattle.ui import Observer, Subject
 from algobattle.team import Matchup, Team, TeamHandler
 from algobattle.problem import Problem
